removed old Force.cpp and renamed force_dev.cpp to Force.cpp
